AMLS-3873 | WIP upversioning service changes
